requirements.txt
gcalcli
Pillow
cairosvg